Used the repository pattern and updated the affected tests
